chore : service1, controller1 삭제 (#96)
