+++ src/main/java/com/rubminds/api/post/dto/PostResponse.java
@@ -28,54 +28,6 @@ public class PostResponse {
 
     }
 
-//    @Getter
-//    @Builder
-//    @NoArgsConstructor(access = AccessLevel.PRIVATE)
-//    @AllArgsConstructor(access = AccessLevel.PRIVATE)
-//    public static class Info {
-//        private Long id;
-//        private PostDto.Writer writer;
-//        private String title;
-//        private String content;
-//        private int headcount;
-//        private String meeting;
-//        private String postsStatus;
-//        private String region;
-//        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
-//        private LocalDateTime createAt;
-//        private List<String> postSkills;
-//        private List<String> customSkills;
-//        private Boolean isLike;
-//        private Long teamId;
-//        private List<PostDto.File> files;
-//        private List<PostDto.File> completeFiles;
-//        private String refLink;
-//        private String completeContent;
-//        private Kinds kinds;
-//
-//        public static PostResponse.Info build(Post post, CustomUserDetails customUserDetails, List<PostFile> postfiles, List<PostFile> completeFiles) {
-//            return Info.builder()
-//                    .id(post.getId())
-//                    .writer(PostDto.Writer.build(post.getWriter()))
-//                    .title(post.getTitle())
-//                    .content(post.getContent())
-//                    .headcount(post.getHeadcount())
-//                    .meeting(post.getMeeting().name())
-//                    .kinds(post.getKinds())
-//                    .postsStatus(post.getPostStatus().name())
-//                    .region(post.getRegion())
-//                    .postSkills(post.getPostSkills().stream().map(postSkill -> postSkill.getSkill().getName()).collect(Collectors.toList()))
-//                    .customSkills(post.getCustomSkills().stream().map(CustomSkill::getName).collect(Collectors.toList()))
-//                    .isLike(post.isLike(customUserDetails))
-//                    .teamId(post.getTeam().getId())
-//                    .files(postfiles.stream().map(PostDto.File::build).collect(Collectors.toList()))
-//                    .completeFiles(completeFiles.stream().map(PostDto.File::build).collect(Collectors.toList()))
-//                    .completeContent(post.getContent())
-//                    .refLink(post.getRefLink())
-//                    .createAt(post.getCreatedAt())
-//                    .build();
-//        }
-//    }
 
     @Getter
     @Builder
