+++ src/rest/PostController.http
@@ -1,4 +1,4 @@
-GET {{apiUrl}}/post/5
+GET {{apiUrl}}/post/1
 Authorization: {{authorizationToken}}
 Content-Type: application/json;charset=UTF-8
